Better validation of blocks & fluids in heat properties recipe
There does not necessarily need to be a mod loaded for the domain part
of the block/fluid resource location; e.g. KubeJS could be used to
register blocks in a custom namespace.  So check for the existence of
the block/fluid in the corresponding registry before do any mod-loaded
checks.

#1236