Add Torus area type
This change adds the Torus area type. P1 is assumed to be in the center
of the torus. When radially symmetric about the Y-axis, the distance
between P1 and P2 in the X/Z plane makes the radius of the torus to the
center of the tube and the distance in Y makes the radius of the tube.

Algorithm and variable naming from
https://en.wikipedia.org/wiki/Torus